Created Main Menu + Help Page
Created 2 new unity scenes, one for the main menu and one for the help page.
Created a new script for link to the buttons on the respected pages.
Styled both scenes.
Help button does work and takes user to help page, for now the other buttons send a console log.
Next step will be to link the start button to the first scene of the game...