Pause menu + first 'test' scene
Linked the start menu play button to load a first level (copy of sample scene for now)
Implemented a pause button that pauses the game when 'esc' key is pressed. Gives the player an option to return to main menu or press 'esc' again to resume game.
Saved pause + menu objects as prefabs for the actual game levels.